+++ tests/index.test.ts
@@ -1,3 +1,5 @@
+import { GetSecretValueCommand, SecretsManagerClient } from '@aws-sdk/client-secrets-manager';
+import { mockClient } from 'aws-sdk-client-mock';
 import { describe, expect, it } from 'vitest';
 
 import {
@@ -52,5 +54,18 @@ describe('aws-sdk-client-mock-vitest', () => {
     'toReceiveAnyCommand',
   ])('should be able to extend with %s', (matcher) => {
     expect(expect('something')).toHaveProperty(matcher);
+  });
+
+  // https://github.com/stschulte/aws-sdk-client-mock-vitest/issues/5
+  it('mocks aws secret manager', async () => {
+    const secretMock = mockClient(SecretsManagerClient);
+    secretMock.on(GetSecretValueCommand).resolves({ SecretString: 'secr3t' });
+
+    const sm = new SecretsManagerClient();
+    const command = new GetSecretValueCommand({ SecretId: 'foo' });
+    const response = await sm.send(command);
+    expect(response.SecretString).toBe('secr3t');
+
+    expect(secretMock).toHaveReceivedCommandWith(GetSecretValueCommand, { SecretId: 'foo' });
   });
 });